Move TestPackedSequenceLogprobs to test_grpo_fast_gpu.py
It belongs with the GPU tests since it requires CUDA + FA2, and it
tests grpo_utils.forward_for_logprobs, not olmo_core.

Made-with: Cursor
